+++ CHANGELOG.md
@@ -6,13 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Computer themes have custom buttons
+
 ### Changed
 
 - Laptops can now be picked up
 
 - ### Contributors for this release
 
-- Fl1s
+- Fl1s, medsal15
 
 ## [1.21.1-1.14.0.1] - 2026-06-01